+++ src/core/hss/various/HSSMultipleSelection.h
@@ -99,7 +99,7 @@ namespace AXR
         const_iterator begin() const;
         iterator end();
         const_iterator end() const;
-        unsigned size() const;
+        size_t size() const;
         void clear();
         bool empty() const;
         HSSSimpleSelection::p front() const;
